requirements.txt
pygame
python-dotenv
supabase
